+++ internal/config/load.go
@@ -32,6 +32,12 @@ func applyEnv(c *Config) error {
 	envStr("WEFT_BASE_DN", &c.BaseDN)
 	envStr("WEFT_ADMIN_UID", &c.AdminUID)
 	envStr("WEFT_ADMIN_DN", &c.AdminDN)
+	envStr("WEFT_CHROOT", &c.Chroot)
+	envStr("WEFT_USER", &c.User)
+	envStr("WEFT_GROUP", &c.Group)
+	if err := envBool("WEFT_SANDBOX", &c.Sandbox); err != nil {
+		return err
+	}
 	envStr("WEFT_LISTEN_ADDR", &c.ListenAddr)
 	envStr("WEFT_CA_CERT_FILE", &c.CACertFile)
 	envStr("WEFT_TLS_CERT_FILE", &c.TLSCertFile)