+++ internal/directory/ldapd/assert_test.go
@@ -1,10 +1,9 @@
 package ldapd
 
 import (
-	"weft/internal/config"
 	"weft/internal/directory"
 )
 
 // Compile-time assertions that the implementation satisfies the interfaces.
-var _ directory.Directory = New(config.Default())
+var _ directory.Directory = (*Directory)(nil)
 var _ directory.Conn = (*conn)(nil)